Slight reorganization of handlers for msw
